sensors/requirements.txt
w1thermsensor
requests
RPi.GPIO
pigpio
gpiozero
smbus2
pimoroni-bme280
prometheus_client
